+++ src/arlg/entity/enemy.cc
@@ -1,5 +1,4 @@
 #include "enemy.h"
-#include "../../config/colDecider.h"
 #include "../../engine/movement/movementComponent.h"
 #include "../../engine/sprite/sprite.h"
 #include "../../engine/util/acceptor.h"
@@ -8,8 +7,8 @@
 
 using std::make_unique;
 
-Enemy::Enemy(int x, int y, unique_ptr<Sprite> s, unique_ptr<MovementComponent> mvt, int health) : Acceptor{x, y, std::move(s), std::move(mvt), make_unique<EnemyCollider>(this), health} {}
+Enemy::Enemy(int x, int y, unique_ptr<Sprite> s, unique_ptr<MovementComponent> mvt, float health) : Acceptor{x, y, std::move(s), std::move(mvt), make_unique<EnemyCollider>(this), health} {}
 
-Enemy::Enemy(int x, int y, unique_ptr<Sprite> s, int health) : Acceptor{x, y, std::move(s), make_unique<EnemyCollider>(this), health} {}
+Enemy::Enemy(int x, int y, unique_ptr<Sprite> s, float health) : Acceptor{x, y, std::move(s), make_unique<EnemyCollider>(this), health} {}
 
 bool Enemy::isEnemy() { return true; }
